+++ NEWS.md
@@ -7,6 +7,8 @@
   `make-orthogonal` can be now restricted to one or more R versions.
   `rim add` now only calls them for the newly installed R version.
 
+* `rim system add-pak` is now implemented.
+
 # rim 0.1.0
 
 First pre-release.